Move public endpoints to bp
