+++ nfdi_search_engine/web/public/__init__.py
@@ -1,0 +1,5 @@
+from flask import Blueprint
+
+bp = Blueprint("public", __name__)
+
+from . import routes
